+++ anvil_consortium_manager/urls.py
@@ -48,6 +48,11 @@ account_patterns = (
             views.AccountLinkVerify.as_view(),
             name="verify",
         ),
+        path(
+            "<uuid:uuid>/add_to_group/",
+            views.GroupAccountMembershipCreateByAccount.as_view(),
+            name="add_to_group",
+        ),
         path("audit/", views.AccountAudit.as_view(), name="audit"),
     ],
     "accounts",
@@ -56,9 +61,19 @@ account_patterns = (
 member_group_patterns = (
     [
         path(
+            "new/",
+            views.GroupGroupMembershipCreateByParent.as_view(),
+            name="new",
+        ),
+        path(
             "<slug:child_group_slug>/",
             views.GroupGroupMembershipDetail.as_view(),
             name="detail",
+        ),
+        path(
+            "<slug:child_group_slug>/new/",
+            views.GroupGroupMembershipCreateByParentChild.as_view(),
+            name="new_by_child",
         ),
         path(
             "<slug:child_group_slug>/delete/",
@@ -72,9 +87,19 @@ member_group_patterns = (
 member_account_patterns = (
     [
         path(
+            "new",
+            views.GroupAccountMembershipCreateByGroup.as_view(),
+            name="new",
+        ),
+        path(
             "<uuid:account_uuid>/",
             views.GroupAccountMembershipDetail.as_view(),
             name="detail",
+        ),
+        path(
+            "<uuid:account_uuid>/new/",
+            views.GroupAccountMembershipCreateByGroupAccount.as_view(),
+            name="new_by_account",
         ),
         path(
             "<uuid:account_uuid>/delete/",
@@ -115,6 +140,11 @@ managed_group_patterns = (
         path("<slug:parent_group_slug>/member_groups/", include(member_group_patterns)),
         path("<slug:group_slug>/member_accounts/", include(member_account_patterns)),
         path("<slug:group_slug>/sharing/", include(managed_group_sharing_patterns)),
+        path(
+            "<slug:group_slug>/add_to_group/",
+            views.GroupGroupMembershipCreateByChild.as_view(),
+            name="add_to_group",
+        ),
     ],
     "managed_groups",
 )